chore: move example to uvu
