+++ package.json
@@ -6,7 +6,7 @@
   "private": true,
   "scripts": {
     "init": "pnpm install",
-    "test:js": "mocha && pnpm --filter arco-design-pro run build",
+    "test:js": "pnpm -r run test",
     "snapshot:js": "mocha --update",
     "test:rust": "cargo test --all -- --nocapture",
     "test:example": "cargo run --example arco_pro",
@@ -21,7 +21,6 @@
     "build:webpack": "cd webpack-examples && node buildAll.js",
     "build:cli": "pnpm --filter @rspack/core... build",
     "build:binding": "pnpm --filter @rspack/binding build",
-    "build_binding_and_run_example": "pnpm build:binding && pnpm --filter @rspack/core run example ",
     "version:snapshot": "changeset version --snapshot",
     "release:snapshot": "changeset publish --tag snapshot --no-git-tag --no-git-checks",
     "release": "changeset publish --no-git-checks",
