+++ packages/rspack-dev-server/package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "build": "rm -rf dist/ && tsc -p tsconfig.json",
     "dev": "tsc -p tsconfig.json -w",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo success"
   },
   "author": "",
   "license": "ISC",
