+++ packages/rspack-plugin-postcss/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "src/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo success"
   },
   "keywords": [],
   "dependencies": {
