+++ packages/rspack/package.json
@@ -25,7 +25,9 @@
     "@types/connect": "3.4.35",
     "uvu": "0.5.6",
     "@rspack/core": "workspace:*",
-    "@types/webpack-sources": "3.2.0"
+    "@types/webpack-sources": "3.2.0",
+    "sinon": "14.0.0",
+    "@types/sinon": "10.0.13"
   },
   "dependencies": {
     "@rspack/binding": "workspace:*",
