+++ webpack-examples/package.json
@@ -5,7 +5,7 @@
   "private": true,
   "main": "build-common.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo success"
   },
   "devDependencies": {
     "@rspack/core": "workspace:*"
